backend/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
google-genai
pypdf
python-multipart
requests
spacy